Wait for all messages being received
Previously, the subscribers sent a DISCONNECT to the server 3 seconds
(hardcoded) after the publishers were done with publishing.

In scenarios like
```
./mqtt-bm-latency -clients 1 -count 1000000 -pubqos 0 -subqos 0
```
the publisher fires all messages at once to the server without waiting
for a response.

In such cases 3 seconds were not enough for all messages being sent from
the server to the subscriber.

This commit ensures that the subscriber sends the DISCONNECT only once
it received all messages.

This assumes that there is no message loss.
Message loss is not expected when testing on the local machine.